+++ example/FamilyHubs.Example/Pages/Examples/FullPages/Radios/Radio-Custom.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FamilyHubs.Example.Pages.Examples.FullPages.RadioCustomModel
+@model RadioCustomModel
 @{
     ViewData["Title"] = "Where do you live?";
 }
