+++ example/FamilyHubs.Example/Pages/Examples/FullPages/Radios/Radio-Inline.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FamilyHubs.Example.Pages.Examples.FullPages.RadioModel
+@model RadioInlineModel
 @{
     ViewData["Title"] = Model.Legend;
 }
