+++ example/FamilyHubs.Example/Pages/Examples/FullPages/Radios/Radio-Inline.cshtml.cs
@@ -4,11 +4,11 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using System.Collections.Immutable;
 
-namespace FamilyHubs.Example.Pages.Examples.FullPages;
+namespace FamilyHubs.Example.Pages.Examples.FullPages.Radios;
 
 public class RadioInlineModel : PageModel, IRadiosPageModel
 {
-    public static Radio[] StaticRadios => new []
+    public static Radio[] StaticRadios => new[]
     {
         new Radio("England", Country.England.ToString()),
         new Radio("Scotland", Country.Scotland.ToString()),
