+++ example/FamilyHubs.Example/Pages/Examples/FullPages/Radios/Radio.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FamilyHubs.Example.Pages.Examples.FullPages.RadioInlineModel
+@model RadioModel
 @{
     ViewData["Title"] = Model.Legend;
 }
