+++ example/FamilyHubs.Example/Pages/Examples/FullPages/Radios/Radio.cshtml.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using System.Collections.Immutable;
 
-namespace FamilyHubs.Example.Pages.Examples.FullPages
+namespace FamilyHubs.Example.Pages.Examples.FullPages.Radios
 {
     public enum Country
     {
@@ -16,7 +16,7 @@ namespace FamilyHubs.Example.Pages.Examples.FullPages
 
     public class RadioModel : PageModel, IRadiosPageModel
     {
-        public static Radio[] StaticRadios => new []
+        public static Radio[] StaticRadios => new[]
         {
             new Radio("England", Country.England.ToString()),
             new Radio("Scotland", Country.Scotland.ToString()),
