+++ example/FamilyHubs.Example/Pages/Examples/FullPages/SingleAutocomplete/SingleAutocomplete.cshtml
@@ -1,0 +1,7 @@
+﻿@page
+@model SingleAutocompleteModel
+@{
+    ViewData["Title"] = "";
+}
+
+<partial name="_SingleAutocompletePage" model="Model" />
