+++ example/FamilyHubs.Example/Pages/Examples/FullPages/SingleAutocomplete/SingleAutocomplete.cshtml.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace FamilyHubs.Example.Pages.Examples.FullPages.SingleAutocomplete
+{
+    public class SingleAutocompleteModel : PageModel
+    {
+        public void OnGet()
+        {
+        }
+    }
+}
