+++ example/FamilyHubs.Example/Pages/Examples/FullPages/SingleTextArea/SingleTextArea.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model FamilyHubs.Example.Pages.Examples.FullPages.SingleTextAreaModel
+@model SingleTextAreaModel
 @{
     ViewData["Title"] = "What are your favourite animals?";
 }
