+++ example/FamilyHubs.Example/Pages/Index.cshtml
@@ -58,13 +58,14 @@
         <h3 class="govuk-heading-s">Full pages:</h3>
 
         <ul class="govuk-list">
-            <li><a href="/Examples/FullPages/SingleTextbox">Single textbox</a></li>
-            <li><a href="/Examples/FullPages/SingleTextboxSeparateHeader">Single textbox (separate header)</a></li>
-            <li><a href="/Examples/FullPages/SingleTextArea">Single text area</a></li>
-            <li><a href="/Examples/FullPages/Radio">Radio</a></li>
-            <li><a href="/Examples/FullPages/Radio-Inline">Inline Radio</a></li>
-            <li><a href="/Examples/FullPages/Radio-Custom">Radio with custom content</a></li>
-            <li><a href="/Examples/FullPages/Checkbox-Custom">Checkbox with custom content</a></li>
+            <li><a href="/Examples/FullPages/SingleAutocomplete/SingleAutocomplete">Single autocomplete</a></li>
+            <li><a href="/Examples/FullPages/SingleTextbox/SingleTextbox">Single textbox</a></li>
+            <li><a href="/Examples/FullPages/SingleTextbox/SingleTextboxSeparateHeader">Single textbox (separate header)</a></li>
+            <li><a href="/Examples/FullPages/SingleTextarea/SingleTextArea">Single text area</a></li>
+            <li><a href="/Examples/FullPages/Radios/Radio">Radio</a></li>
+            <li><a href="/Examples/FullPages/Radios/Radio-Inline">Inline Radio</a></li>
+            <li><a href="/Examples/FullPages/Radios/Radio-Custom">Radio with custom content</a></li>
+            <li><a href="/Examples/FullPages/Checkboxes/Checkbox-Custom">Checkbox with custom content</a></li>
         </ul>
 
         <h3 class="govuk-heading-s">Header:</h3>
