+++ fh-web-components.sln
@@ -45,6 +45,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "components", "components", 
 		src\familyhubs-frontend\scripts\components\add-another.ts = src\familyhubs-frontend\scripts\components\add-another.ts
 		src\familyhubs-frontend\scripts\components\analytics.ts = src\familyhubs-frontend\scripts\components\analytics.ts
 		src\familyhubs-frontend\scripts\components\back-links.ts = src\familyhubs-frontend\scripts\components\back-links.ts
+		src\familyhubs-frontend\scripts\components\clarity.ts = src\familyhubs-frontend\scripts\components\clarity.ts
 		src\familyhubs-frontend\scripts\components\cookie-banner.ts = src\familyhubs-frontend\scripts\components\cookie-banner.ts
 		src\familyhubs-frontend\scripts\components\cookie-functions.ts = src\familyhubs-frontend\scripts\components\cookie-functions.ts
 		src\familyhubs-frontend\scripts\components\cookies-page.ts = src\familyhubs-frontend\scripts\components\cookies-page.ts
@@ -52,7 +53,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "components", "components", 
 		src\familyhubs-frontend\scripts\components\open-close-button.ts = src\familyhubs-frontend\scripts\components\open-close-button.ts
 		src\familyhubs-frontend\scripts\components\postcode.ts = src\familyhubs-frontend\scripts\components\postcode.ts
 		src\familyhubs-frontend\scripts\components\visibility-toggle.ts = src\familyhubs-frontend\scripts\components\visibility-toggle.ts
-		src\familyhubs-frontend\scripts\components\clarity.ts = src\familyhubs-frontend\scripts\components\clarity.ts
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "layout", "layout", "{788DA990-BC17-4572-B087-042C91FF99A0}"
