+++ src/FamilyHubs.SharedKernel.Razor/Pages/Shared/_SingleAutocompletePage.cshtml
@@ -16,7 +16,7 @@
                         <!---Select local authority--->
                     </label>
                     <select class="govuk-select" id="local-authority" name="la">
-                        <option value="@local_authorityModel.NoSelectionId" selected disabled></option>
+                        <option value="-1" selected disabled></option>
                         @foreach (var organisation in Model.Organisations)
                         {
                             <option value="@organisation.Id">@organisation.Name</option>
